modelApp/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.0
python-dateutil==2.8.0
pytorch-ignite==0.2.1
pytz==2019.2
requests==2.22.0
six==1.12.0
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
urllib3==1.25.6
Werkzeug==0.16.0
